requirements.txt
fastapi[standard]
requests
langchain-openai
langchain-core
openai
tiktoken
python-dotenv
supabase